Make uhdm-dump a toplevel utility binary.
Dumping uhdm is a critical part of working with uhdm, so
let's make this a toplevel binary, installed in $PREFIX/bin/uhdm-dump.

Previously, the testing binary used in tests/dump.cpp was used for that,
but let's keep that just a simple round-trip test.

Additional context: not only makes it sense in this context, but also
in Surelog, in which we'd like to disable building UDHM tests, but still
need a binary to dump uhdm.

Signed-off-by: Henner Zeller <h.zeller@acm.org>