vpi_user: handle some properties from the base-class once.
The base class offers a few properties, such as VpiLineNo() or
VpiFile() that can be handled once in vpi_get() and vpi_get_str(),
so that the remaining function body only has to worry about the
'non-standard' properties.

 * Changed vpi_get64()/vpi_get() to handle VpiLineNo() and VpiType().
 * All other properties are output in a similar way as before,
   but in a slightly easier human-browsable switch/case form.
 * Since get64 and get are the same right now, just one one function
   call the other so that we don't duplicate code.
 * vpi_get_str handles VpiFile, VpiName and VpiDefName from the
   base-class, then deals with remaining 'manually'.

There are still a lot of properties that have to be handled
manually, but future work could consider having property-specific
interfaces ('VpiFullNamePropertyContainer') that could further
reduce that.

Overall, this reduces the number of branches the compiler has to
encode and the runtime has to consider a lot. Dumping BlackParrot2
got 20% faster.

Signed-off-by: Henner Zeller <h.zeller@acm.org>