Only write generated files if they actually changed.
This reduces churn of what the compiler has to re-compile a lot
when doing small changes to code generation and object models:
only files that don't exist or whose output actually change
are be overwritten.

Tested:
  touch model/*.yaml ; make
The tcl script is run, but it does not generate any changed artifacts,
so no files are recompiled.

Signed-off-by: Henner Zeller <h.zeller@acm.org>