Install all uhdm related headers in flat $PREFIX/include/uhdm/*.h
This removes the arbitrary distinction between relatively
meaningless names header/ and include/

Adapt all code that used the previous paths in their includes
to point to the new location.

Install location is presumed to be a 'system location', i.e.
the includes are used in angled brackets,
such as #include <uhdm/uhdm.h>

This likely cause one-time disruption for current users, but
currently the user-base is small enough to be a limited
disruption.

While at it, install also the UHDM.capnp in $PREFIX/lib/uhdm
for other tools using different languages to pick up just the
schema.

Fixes #484

Signed-off-by: Henner Zeller <h.zeller@acm.org>