Modularization PAss 2
Take model class hierarchy into account when generating code for
vpi_listener, vpi_user and vpi_visitor. Also, the order of
listen/visit callbacks are in the same order as the model's
definition.

NOTE: This implementation changes the order of member processing.
Previous implementation used to visit/listen member of the most
concrete class before doing it for base. This implementation does
it in reverse i.e. the base classes are processed before the most
derived class.