Use precompiled headers to improve build times.
Initial numbers look very promising with Windows showing
over 70% improvement on build times.

Gotcha - Certain files have different optimization flags at
the moment and those differ from the ones the precompiled header
is built with. This shows up as an issue with clang on MacOS but
none of the other platforms. Resolution is to disable use of
precompiled headers for those certain files (unfortunately, those
are ones that can benefit the most with use of precompiled header).