Using std::deque<> as an allocation space for factories.
This allows to have pre-sized buffers and easily de-allocated objects.
Since malloc allocators already do something similar, this is not
creating huge saving (about 70 MiB RAM saving for Rsd, 5-10% CPU time)

This particular implementation will increase the compile time by
a lot as we need to pull in all the classes from uhdm/uhdm.h (as
we need to know the size in the factory), so it would need some
refinement (e.g. some PIMPL-style) to be usable.

So leaving it here as draft if someone wants to experiment with
it.