Update to callback logic in UhdmListener
In previous implementation, specific vpiXXX listen invocations were
skipped. This has been found to be a problem since there are cases where
there is only one such edge in the tree and that is being skipped.

Instead, let the client code deal with duplicate callbacks. The visit
container will track all objects that are already stepped-in to avoid
infinite recursion.