Issue #3212: Remove duplicate string table implementation
Move Surelog/SymbolTable to UHDM as SymbolFactory to replace its own
implementation. Also, pulled in SymbolId implementation.

Updated code generation scripts to use SymbolId instead of
SymbolFactory::ID.

IMPORTANT NOTE:
One important difference between UHDM's and Surelog's use of SymbolTable
is that UHDM resolves an invalid SymbolId to an empty string but
Surelog's expects the BAD_SYMBOL. So, the API in the implementation
can't be the same. UHDM has its own set of APIs and Surelog it's own.