Allow to use native capnproto.
By default, keeps using third_party/ capnproto, but allows to
use the capnroto installed on the system by setting
UHDM_USE_HOST_CAPNP=On
This helps UHDM to be packed in distributions.

(So wht do we keep third_party/ ? On Windows, there is no common
package manager, so projects there rely more on things to be pulled
in together; see discussion in #491).

Fixes: #483