Don't use -DDEBUG to change layout of SymbolId
Since -DDEBUG is commonly used, it is dangerous to change the
layout of an object just because of that - the project using
UHDM will now accidentally change the size that is not compatible
anymore with the library.

(This might not have detected before as UHDM typically was
compiled alongside Surelog with the same flags, but 3rd party
applications just using UHDM will break).

Fixes #872