pkg-config: base inc dir; choose libs depending on USE_HOST_CAPNP
In all our examples, we use `#include <uhdm/foo.h>`, so the base include
directory should be $PREFIX/include, not $PREFIX/include/uhdm.

The libraries to link depends on of iwe used USE_HOST_CAPNP. If we
compiled and installed from third_party/, we need to include the
linking of the libraries from there, otherwise we need to require to
use the pkg-config from capnp.