Fix serialization bug - Bad object index
The bookkeeping for map Serializer::allIds was wrong. When objects were
deleted, allIds_ continued to hold the old index. The collection needed
to be cleared at the start of Serializer::Save and rebuilt again, which
wasn't happening. Also, the use of incrId and objId was inconsistent.

Simplified the logic to remove the bookkeeping all together since the
map doesn't need to be persistent (needed only during save).

Found at least one another bug in the process - Logic assumed that
Serialzer::AllObjects returns a full list of objects but that was true
only if Serializer::Save was called. Serializer::Restore never populated
this collected, since restore logic depends on Serializer::GetObject for
binding.