Fix object-id issue with save-load-edit-save-load cycle
When saving a db, the objects save the assigned id and reload the same.
However, if new objects are created after loading the db, logic ends up
creating objects with index starting at zero. Result is multiple objects
end up with the same object-id - which should never happen since these
are meant to be unique and persistent.

Resolution is the save/load the last usuable object-id so it is always a
monotonically increasing number.