Add IWYU pragmas in header meant to export other headers.
clang-tidy usually would complain if not a direct header for a particular
symbol is included.

The uhdm.h header is meant as a 'catchall' header that provides all
the symbols in all the headers is includes. So mark them as such so that
toolings understand that this is the intention.

A common way is to use the IWYU pragmas for that:
https://github.com/include-what-you-use/include-what-you-use/blob/master/docs/IWYUPragmas.md#iwyu-pragma-begin_exportsend_exports